+++ config/autoload/doctrine.global.php
@@ -20,6 +20,7 @@ return [
             'orm_default' => [
                 'doctrine_mapping_types' => [
                     UuidBinaryType::NAME => UuidBinaryType::BINARY,
+                    'enum' => 'string',
                 ],
             ],
         ],
@@ -32,7 +33,7 @@ return [
                 ],
             ],
         ],
-        'migrations_configuration' => [
+        'migrations' => [
             'orm_default' => [
                 'directory' => 'data/migrations',
                 'name'      => 'FactorioItemBrowser ExportQueue Database Migrations',
@@ -41,6 +42,7 @@ return [
             ],
         ],
         'types' => [
+            Doctrine\Type\JobPriorityType::NAME => Doctrine\Type\JobPriorityType::class,
             Doctrine\Type\JobStatusType::NAME => Doctrine\Type\JobStatusType::class,
             UuidBinaryType::NAME => UuidBinaryType::class,
         ],