+++ src/Doctrine/Type/JobStatusType.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace FactorioItemBrowser\ExportQueue\Server\Doctrine\Type;
 
-use Doctrine\DBAL\Platforms\AbstractPlatform;
-use Doctrine\DBAL\Types\Type;
 use FactorioItemBrowser\ExportQueue\Client\Constant\JobStatus;
 
 /**
@@ -14,7 +12,7 @@ use FactorioItemBrowser\ExportQueue\Client\Constant\JobStatus;
  * @author BluePsyduck <bluepsyduck@gmx.com>
  * @license http://opensource.org/licenses/GPL-3.0 GPL v3
  */
-class JobStatusType extends Type
+class JobStatusType extends AbstractEnumType
 {
     /**
      * The name of the Doctrine type.
@@ -24,7 +22,7 @@ class JobStatusType extends Type
     /**
      * The values of the enum.
      */
-    protected const VALUES = [
+    public const VALUES = [
         JobStatus::QUEUED,
         JobStatus::DOWNLOADING,
         JobStatus::PROCESSING,
@@ -34,31 +32,4 @@ class JobStatusType extends Type
         JobStatus::DONE,
         JobStatus::ERROR,
     ];
-
-    /**
-     * Gets the SQL declaration snippet for a field of this type.
-     *
-     * @param mixed[] $fieldDeclaration The field declaration.
-     * @param AbstractPlatform $platform The currently used database platform.
-     *
-     * @return string
-     */
-    public function getSQLDeclaration(array $fieldDeclaration, AbstractPlatform $platform)
-    {
-        $values = implode(',', array_map(function (string $value) use ($platform): string {
-            return $platform->quoteStringLiteral($value);
-        }, self::VALUES));
-
-        return sprintf('ENUM(%s)', $values);
-    }
-
-    /**
-     * Gets the name of this type.
-     *
-     * @return string
-     */
-    public function getName()
-    {
-        return self::NAME;
-    }
 }